requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
colorlog==6.7.0
frozenlist==1.3.3
hikari==2.0.0.dev118
hikari-lightbulb==2.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
valorant==1.0.4
yarl==1.9.2